restructured CLI layout
